TensorRT refactoring
