+++ src/Detector/tensorrt_yolo/YoloONNX.cpp
@@ -495,823 +495,3 @@ bool YoloONNX::VerifyOutputAspectRatio(size_t imgIdx, std::vector<tensor_rt::Res
 
     return !outputs.empty();
 }
-
-///
-/// \brief YoloONNX::GetResult
-/// \param output
-/// \return
-///
-std::vector<tensor_rt::Result> YoloONNX::GetResult(size_t imgIdx, int /*keep_topk*/, const std::vector<float*>& outputs, cv::Size frameSize)
-{
-    std::vector<tensor_rt::Result> resBoxes;
-    if (m_params.m_netType == tensor_rt::ModelType::YOLOV7Mask || m_params.m_netType == tensor_rt::ModelType::YOLOV8Mask)
-        ProcessMaskOutput(imgIdx, outputs, frameSize, resBoxes);
-    else
-        ProcessBBoxesOutput(imgIdx, outputs, frameSize, resBoxes);
-    return resBoxes;
-}
-
-///
-/// \brief YoloONNX::ProcessMaskOutput
-/// \param output
-/// \return
-///
-void YoloONNX::ProcessMaskOutput(size_t imgIdx, const std::vector<float*>& outputs, cv::Size frameSize, std::vector<tensor_rt::Result>& resBoxes)
-{
-    if (m_params.m_netType == tensor_rt::ModelType::YOLOV7Mask)
-    {
-        const float fw = static_cast<float>(frameSize.width) / static_cast<float>(m_inputDims.d[3]);
-        const float fh = static_cast<float>(frameSize.height) / static_cast<float>(m_inputDims.d[2]);
-
-        size_t outInd = (outputs.size() == 0) ? 0 : 1;
-        size_t segInd = (outputs.size() == 0) ? 1 : 0;
-
-        auto output = outputs[0];
-
-        //0: name: images, size : 1x3x640x640
-        //1 : name : 516, size : 1x32x160x160
-        //2 : name : onnx::Slice_542, size : 1x3x80x80x117
-        //3 : name : onnx::Slice_710, size : 1x3x40x40x117
-        //4 : name : onnx::Slice_878, size : 1x3x20x20x117
-        //5 : name : output, size : 1x25200x117
-        // 25200 = 3x80x80 + 3x40x40 + 3x20x20
-        // 117 = x, y, w, h, c, 80 classes, 32 seg ancors
-        // 80 * 8 = 640, 40 * 16 = 640, 20 * 32 = 640
-
-        size_t ncInd = 2;
-        size_t lenInd = 1;
-        if (m_outpuDims[outInd].nbDims == 2)
-        {
-            ncInd = 1;
-            lenInd = 0;
-        }
-        int nc = m_outpuDims[outInd].d[ncInd] - 5 - 32;
-        size_t len = static_cast<size_t>(m_outpuDims[outInd].d[lenInd]) / m_params.explicitBatchSize;
-        //auto Volume = [](const nvinfer1::Dims& d)
-        //{
-        //    return std::accumulate(d.d, d.d + d.nbDims, 1, std::multiplies<int>());
-        //};
-        auto volume = len * m_outpuDims[outInd].d[ncInd]; // Volume(m_outpuDims[0]);
-        output += volume * imgIdx;
-        //std::cout << "len = " << len << ", nc = " << nc << ", m_params.confThreshold = " << m_params.confThreshold << ", volume = " << volume << std::endl;
-
-#if 1
-        int segWidth = 160;
-        int segHeight = 160;
-        int segChannels = 32;
-
-        if (outputs.size() > 1)
-        {
-            //std::cout << "516 nbDims: " << m_outpuDims[segInd].nbDims << ", ";
-            //for (size_t i = 0; i < m_outpuDims[segInd].nbDims; ++i)
-            //{
-            //    std::cout << m_outpuDims[segInd].d[i];
-            //    if (i + 1 != m_outpuDims[segInd].nbDims)
-            //        std::cout << "x";
-            //}
-            //std::cout << std::endl;
-//
-            //std::cout << "output nbDims: " << m_outpuDims[outInd].nbDims << ", ";
-            //for (size_t i = 0; i < m_outpuDims[outInd].nbDims; ++i)
-            //{
-            //    std::cout << m_outpuDims[outInd].d[i];
-            //    if (i + 1 != m_outpuDims[outInd].nbDims)
-            //        std::cout << "x";
-            //}
-            //std::cout << std::endl;
-
-            segChannels = m_outpuDims[segInd].d[1];
-            segWidth = m_outpuDims[segInd].d[2];
-            segHeight = m_outpuDims[segInd].d[3];
-        }
-        cv::Mat maskProposals;
-        std::vector<std::vector<float>> picked_proposals;
-        int net_width = nc + 5 + segChannels;
-#endif
-
-        std::vector<int> classIds;
-        std::vector<float> confidences;
-        std::vector<cv::Rect> rectBoxes;
-        classIds.reserve(len);
-        confidences.reserve(len);
-        rectBoxes.reserve(len);
-
-        for (size_t i = 0; i < len; ++i)
-        {
-            // Box
-            size_t k = i * (nc + 5 + 32);
-            float object_conf = output[k + 4];
-
-            if (i == 0)
-            {
-                //std::cout << "without nms: mem" << i << ": ";
-                //for (size_t ii = 0; ii < nc; ++ii)
-                //{
-                //    std::cout << output[k + ii] << " ";
-                //}
-                //std::cout << ";" << std::endl;
-                //for (size_t ii = nc; ii < nc + 4; ++ii)
-                //{
-                //    std::cout << output[k + ii] << " ";
-                //}
-                //std::cout << ";" << std::endl;
-                //for (size_t ii = nc + 4; ii < nc + 4 + 32; ++ii)
-                //{
-                //    std::cout << output[k + ii] << " ";
-                //}
-                //std::cout << ";" << std::endl;
-            }
-
-            if (object_conf >= m_params.confThreshold)
-            {
-                // (center x, center y, width, height) to (x, y, w, h)
-                float x = fw * (output[k] - output[k + 2] / 2);
-                float y = fh * (output[k + 1] - output[k + 3] / 2);
-                float width = fw * output[k + 2];
-                float height = fh * output[k + 3];
-
-                // Classes
-                float class_conf = output[k + 5];
-                int classId = 0;
-
-                for (int j = 1; j < nc; ++j)
-                {
-                    if (class_conf < output[k + 5 + j])
-                    {
-                        classId = j;
-                        class_conf = output[k + 5 + j];
-                    }
-                }
-
-                class_conf *= object_conf;
-
-                //if (i == 0)
-                //    std::cout << i << ": object_conf = " << object_conf << ", class_conf = " << class_conf << ", classId = " << classId << ", rect = " << cv::Rect(cvRound(x), cvRound(y), cvRound(width), cvRound(height)) << std::endl;
-
-                classIds.push_back(classId);
-                confidences.push_back(class_conf);
-                rectBoxes.emplace_back(cvRound(x), cvRound(y), cvRound(width), cvRound(height));
-
-                std::vector<float> temp_proto(output + k + 5 + nc, output + k + net_width);
-                picked_proposals.push_back(temp_proto);
-
-                if (rectBoxes.size() >= CV_CN_MAX - 1)
-                    break;
-            }
-        }
-
-        // Non-maximum suppression to eliminate redudant overlapping boxes
-        std::vector<int> indices;
-        cv::dnn::NMSBoxes(rectBoxes, confidences, m_params.confThreshold, m_params.nmsThreshold, indices);
-        resBoxes.reserve(indices.size());
-
-        for (size_t bi = 0; bi < indices.size(); ++bi)
-        {
-            resBoxes.emplace_back(classIds[indices[bi]], confidences[indices[bi]], Clamp(rectBoxes[indices[bi]], frameSize));
-            maskProposals.push_back(cv::Mat(picked_proposals[indices[bi]]).t());
-        }
-
-        if (!maskProposals.empty())
-        {
-            // Mask processing
-            const float* pdata = outputs[1];
-            std::vector<float> maskFloat(pdata, pdata + segChannels * segWidth * segHeight);
-
-            int INPUT_W = m_inputDims.d[3];
-            int INPUT_H = m_inputDims.d[2];
-            static constexpr float MASK_THRESHOLD = 0.5;
-
-            cv::Mat mask_protos = cv::Mat(maskFloat);
-            cv::Mat protos = mask_protos.reshape(0, { segChannels, segWidth * segHeight });
-
-            cv::Mat matmulRes = (maskProposals * protos).t();//n*32 32*25600
-            cv::Mat masks = matmulRes.reshape(static_cast<int>(resBoxes.size()), { segWidth, segHeight });
-            std::vector<cv::Mat> maskChannels;
-            split(masks, maskChannels);
-            for (size_t i = 0; i < resBoxes.size(); ++i)
-            {
-                cv::Mat dest;
-                cv::Mat mask;
-                //sigmoid
-                cv::exp(-maskChannels[i], dest);
-                dest = 1.0 / (1.0 + dest);//160*160
-
-                int padw = 0;
-                int padh = 0;
-                cv::Rect roi(int((float)padw / INPUT_W * segWidth), int((float)padh / INPUT_H * segHeight), int(segWidth - padw / 2), int(segHeight - padh / 2));
-                dest = dest(roi);
-
-                cv::resize(dest, mask, frameSize, cv::INTER_NEAREST);
-
-                resBoxes[i].m_boxMask = mask(resBoxes[i].m_brect) > MASK_THRESHOLD;
-
-#if 0
-                static int globalObjInd = 0;
-                SaveMat(resBoxes[i].m_boxMask, std::to_string(globalObjInd++), ".png", "tmp", true);
-#endif
-
-                std::vector<std::vector<cv::Point>> contours;
-                std::vector<cv::Vec4i> hierarchy;
-#if (CV_VERSION_MAJOR < 4)
-                cv::findContours(resBoxes[i].m_boxMask, contours, hierarchy, CV_RETR_EXTERNAL, CV_CHAIN_APPROX_SIMPLE, cv::Point());
-#else
-                cv::findContours(resBoxes[i].m_boxMask, contours, hierarchy, cv::RETR_EXTERNAL, cv::CHAIN_APPROX_SIMPLE, cv::Point());
-#endif
-                for (const auto& contour : contours)
-                {
-                    cv::Rect br = cv::boundingRect(contour);
-
-                    if (br.width >= 4 &&
-                        br.height >= 4)
-                    {
-                        cv::RotatedRect rr = (contour.size() < 5) ? cv::minAreaRect(contour) : cv::fitEllipse(contour);
-
-                        br.x += resBoxes[i].m_brect.x;
-                        br.y += resBoxes[i].m_brect.y;
-                        rr.center.x += resBoxes[i].m_brect.x;
-                        rr.center.y += resBoxes[i].m_brect.y;
-
-                        //std::cout << "rr: " << rr.center << ", " << rr.angle << ", " << rr.size << std::endl;
-
-                        if (resBoxes[i].m_boxMask.size() != br.size())
-                        {
-                            br.width = resBoxes[i].m_boxMask.cols;
-                            br.height = resBoxes[i].m_boxMask.rows;
-                            if (br.x + br.width >= frameSize.width)
-                                br.x = frameSize.width - br.width;
-                            if (br.y + br.height >= frameSize.height)
-                                br.y = frameSize.height - br.height;
-                        }
-
-                        resBoxes[i].m_brect = br;
-                        resBoxes[i].m_rrect = rr;
-
-                        break;
-                    }
-                }
-            }
-        }
-    }
-    else if (m_params.m_netType == tensor_rt::ModelType::YOLOV8Mask)
-    {
-        const float fw = static_cast<float>(frameSize.width) / static_cast<float>(m_inputDims.d[3]);
-        const float fh = static_cast<float>(frameSize.height) / static_cast<float>(m_inputDims.d[2]);
-
-        size_t outInd = (outputs.size() == 0) ? 0 : 1;
-        size_t segInd = (outputs.size() == 0) ? 1 : 0;
-
-        auto output = outputs[0];
-
-        //std::cout << "output[1] mem:\n";
-        //auto output1 = outputs[1];
-        //for (size_t ii = 0; ii < 100; ++ii)
-        //{
-        //    std::cout << ii << ": ";
-        //    for (size_t jj = 0; jj < 20; ++jj)
-        //    {
-        //        std::cout << output1[ii * 20 + jj] << " ";
-        //    }
-        //    std::cout << ";" << std::endl;
-        //}
-        //std::cout << ";" << std::endl;
-
-        //0: name: images, size: 1x3x640x640
-        //1: name: output1, size: 1x32x160x160
-        //2: name: output0, size: 1x116x8400
-        // 25200 = 3x80x80 + 3x40x40 + 3x20x20
-        // 116 = x, y, w, h, 80 classes, 32 seg ancors
-        // 80 * 8 = 640, 40 * 16 = 640, 20 * 32 = 640
-
-        size_t ncInd = 1;
-        size_t lenInd = 2;
-        int nc = m_outpuDims[outInd].d[ncInd] - 4 - 32;
-        int dimensions = nc + 32 + 4;
-        size_t len = static_cast<size_t>(m_outpuDims[outInd].d[lenInd]) / m_params.explicitBatchSize;
-        //auto Volume = [](const nvinfer1::Dims& d)
-        //{
-        //    return std::accumulate(d.d, d.d + d.nbDims, 1, std::multiplies<int>());
-        //};
-        auto volume = len * m_outpuDims[outInd].d[ncInd]; // Volume(m_outpuDims[0]);
-        output += volume * imgIdx;
-        //std::cout << "len = " << len << ", nc = " << nc << ", m_params.confThreshold = " << m_params.confThreshold << ", volume = " << volume << std::endl;
-
-        cv::Mat rawMemory(1, dimensions * static_cast<int>(len), CV_32FC1, output);
-        rawMemory = rawMemory.reshape(1, dimensions);
-        cv::transpose(rawMemory, rawMemory);
-        output = (float*)rawMemory.data;
-
-        //std::cout << "output[0] mem:\n";
-        //for (size_t ii = 0; ii < 100; ++ii)
-        //{
-        //    std::cout << ii << ": ";
-        //    for (size_t jj = 0; jj < 20; ++jj)
-        //    {
-        //        std::cout << output[ii * 20 + jj] << " ";
-        //    }
-        //    std::cout << ";" << std::endl;
-        //}
-        //std::cout << ";" << std::endl;
-
-#if 1
-        int segWidth = 160;
-        int segHeight = 160;
-        int segChannels = 32;
-
-        if (outputs.size() > 1)
-        {
-            //std::cout << "output1 nbDims: " << m_outpuDims[segInd].nbDims << ", ";
-            //for (size_t i = 0; i < m_outpuDims[segInd].nbDims; ++i)
-            //{
-            //    std::cout << m_outpuDims[segInd].d[i];
-            //    if (i + 1 != m_outpuDims[segInd].nbDims)
-            //        std::cout << "x";
-            //}
-            //std::cout << std::endl;
-            //std::cout << "output nbDims: " << m_outpuDims[outInd].nbDims << ", ";
-            //for (size_t i = 0; i < m_outpuDims[outInd].nbDims; ++i)
-            //{
-            //    std::cout << m_outpuDims[outInd].d[i];
-            //    if (i + 1 != m_outpuDims[outInd].nbDims)
-            //        std::cout << "x";
-            //}
-            //std::cout << std::endl;
-
-            segChannels = m_outpuDims[segInd].d[1];
-            segWidth = m_outpuDims[segInd].d[2];
-            segHeight = m_outpuDims[segInd].d[3];
-        }
-        cv::Mat maskProposals;
-        std::vector<std::vector<float>> picked_proposals;
-        int net_width = nc + 4 + segChannels;
-#endif
-
-        std::vector<int> classIds;
-        std::vector<float> confidences;
-        std::vector<cv::Rect> rectBoxes;
-        classIds.reserve(len);
-        confidences.reserve(len);
-        rectBoxes.reserve(len);
-
-        for (size_t i = 0; i < len; ++i)
-        {
-            // Box
-            size_t k = i * (nc + 4 + 32);
-
-            int classId = -1;
-            float objectConf = 0.f;
-            for (int j = 0; j < nc; ++j)
-            {
-                const float classConf = output[k + 4 + j];
-                if (classConf > objectConf)
-                {
-                    classId = j;
-                    objectConf = classConf;
-                }
-            }
-
-            //if (i == 0)
-            //{
-            //    std::cout << "without nms: mem" << i << ": ";
-            //    for (size_t ii = 0; ii < 4; ++ii)
-            //    {
-            //        std::cout << output[k + ii] << " ";
-            //    }
-            //    std::cout << ";" << std::endl;
-            //    for (size_t ii = 4; ii < nc + 4; ++ii)
-            //    {
-            //        std::cout << output[k + ii] << " ";
-            //    }
-            //    std::cout << ";" << std::endl;
-            //    for (size_t ii = nc + 4; ii < nc + 4 + 32; ++ii)
-            //    {
-            //        std::cout << output[k + ii] << " ";
-            //    }
-            //    std::cout << ";" << std::endl;
-            //}
-
-            if (objectConf >= m_params.confThreshold)
-            {
-                // (center x, center y, width, height) to (x, y, w, h)
-                float x = fw * (output[k] - output[k + 2] / 2);
-                float y = fh * (output[k + 1] - output[k + 3] / 2);
-                float width = fw * output[k + 2];
-                float height = fh * output[k + 3];
-
-                //auto ClampToFrame = [](float& v, float& size, int hi) -> int
-                //{
-                //    int res = 0;
-//
-                //    if (size < 1)
-                //        size = 0;
-//
-                //    if (v < 0)
-                //    {
-                //        res = v;
-                //        v = 0;
-                //        return res;
-                //    }
-                //    else if (v + size > hi - 1)
-                //    {
-                //        res = v;
-                //        v = hi - 1 - size;
-                //        if (v < 0)
-                //        {
-                //            size += v;
-                //            v = 0;
-                //        }
-                //        res -= v;
-                //        return res;
-                //    }
-                //    return res;
-                //};
-                //ClampToFrame(x, width, frameSize.width);
-                //ClampToFrame(y, height, frameSize.height);
-
-                //if (i == 0)
-                //	std::cout << i << ": object_conf = " << object_conf << ", class_conf = " << class_conf << ", classId = " << classId << ", rect = " << cv::Rect(cvRound(x), cvRound(y), cvRound(width), cvRound(height)) << std::endl;
-
-                if (width > 4 && height > 4)
-                {
-                    classIds.push_back(classId);
-                    confidences.push_back(objectConf);
-                    rectBoxes.emplace_back(cvRound(x), cvRound(y), cvRound(width), cvRound(height));
-
-                    std::vector<float> temp_proto(output + k + 4 + nc, output + k + net_width);
-                    picked_proposals.push_back(temp_proto);
-                }
-            }
-        }
-
-        // Non-maximum suppression to eliminate redudant overlapping boxes
-        std::vector<int> indices;
-        cv::dnn::NMSBoxes(rectBoxes, confidences, m_params.confThreshold, m_params.nmsThreshold, indices);
-        resBoxes.reserve(indices.size());
-
-        for (size_t bi = 0; bi < indices.size(); ++bi)
-        {
-            resBoxes.emplace_back(classIds[indices[bi]], confidences[indices[bi]], Clamp(rectBoxes[indices[bi]], frameSize));
-            maskProposals.push_back(cv::Mat(picked_proposals[indices[bi]]).t());
-        }
-
-        if (!maskProposals.empty())
-        {
-            // Mask processing
-            const float* pdata = outputs[1];
-            std::vector<float> maskFloat(pdata, pdata + segChannels * segWidth * segHeight);
-
-            int INPUT_W = m_inputDims.d[3];
-            int INPUT_H = m_inputDims.d[2];
-            static constexpr float MASK_THRESHOLD = 0.5;
-
-            cv::Mat mask_protos = cv::Mat(maskFloat);
-            cv::Mat protos = mask_protos.reshape(0, { segChannels, segWidth * segHeight });
-
-            cv::Mat matmulRes = (maskProposals * protos).t();//n*32 32*25600
-            cv::Mat masks = matmulRes.reshape(static_cast<int>(resBoxes.size()), { segWidth, segHeight });
-            std::vector<cv::Mat> maskChannels;
-            split(masks, maskChannels);
-            for (size_t i = 0; i < resBoxes.size(); ++i)
-            {
-                cv::Mat dest;
-                cv::Mat mask;
-                //sigmoid
-                cv::exp(-maskChannels[i], dest);
-                dest = 1.0 / (1.0 + dest);//160*160
-
-                int padw = 0;
-                int padh = 0;
-                cv::Rect roi(int((float)padw / INPUT_W * segWidth), int((float)padh / INPUT_H * segHeight), int(segWidth - padw / 2), int(segHeight - padh / 2));
-                dest = dest(roi);
-
-                cv::resize(dest, mask, frameSize, cv::INTER_NEAREST);
-
-                resBoxes[i].m_boxMask = mask(resBoxes[i].m_brect) > MASK_THRESHOLD;
-
-#if 0
-                static int globalObjInd = 0;
-                SaveMat(resBoxes[i].m_boxMask, std::to_string(globalObjInd++), ".png", "tmp", true);
-#endif
-
-                std::vector<std::vector<cv::Point>> contours;
-                std::vector<cv::Vec4i> hierarchy;
-#if (CV_VERSION_MAJOR < 4)
-                cv::findContours(resBoxes[i].m_boxMask, contours, hierarchy, CV_RETR_EXTERNAL, CV_CHAIN_APPROX_SIMPLE, cv::Point());
-#else
-                cv::findContours(resBoxes[i].m_boxMask, contours, hierarchy, cv::RETR_EXTERNAL, cv::CHAIN_APPROX_SIMPLE, cv::Point());
-#endif
-                for (const auto& contour : contours)
-                {
-                    cv::Rect br = cv::boundingRect(contour);
-
-                    if (br.width >= 4 &&
-                        br.height >= 4)
-                    {
-                        cv::RotatedRect rr = (contour.size() < 5) ? cv::minAreaRect(contour) : cv::fitEllipse(contour);
-
-                        br.x += resBoxes[i].m_brect.x;
-                        br.y += resBoxes[i].m_brect.y;
-                        rr.center.x += resBoxes[i].m_brect.x;
-                        rr.center.y += resBoxes[i].m_brect.y;
-
-                        //std::cout << "rr: " << rr.center << ", " << rr.angle << ", " << rr.size << std::endl;
-
-                        if (resBoxes[i].m_boxMask.size() != br.size())
-                        {
-                            br.width = resBoxes[i].m_boxMask.cols;
-                            br.height = resBoxes[i].m_boxMask.rows;
-                            if (br.x + br.width >= frameSize.width)
-                                br.x = frameSize.width - br.width;
-                            if (br.y + br.height >= frameSize.height)
-                                br.y = frameSize.height - br.height;
-                        }
-
-                        resBoxes[i].m_brect = br;
-                        resBoxes[i].m_rrect = rr;
-
-                        break;
-                    }
-                }
-            }
-        }
-    }
-    else
-    {
-        assert(0);
-    }
-}
-
-///
-/// \brief YoloONNX::ProcessBBoxesOutput
-/// \param output
-/// \return
-///
-void YoloONNX::ProcessBBoxesOutput(size_t imgIdx, const std::vector<float*>& outputs, cv::Size frameSize, std::vector<tensor_rt::Result>& resBoxes)
-{
-    if (outputs.size() == 4)
-    {
-        auto dets = reinterpret_cast<int*>(outputs[0]);
-        auto boxes = outputs[1];
-        auto scores = outputs[2];
-        auto classes = reinterpret_cast<int*>(outputs[3]);
-
-        int objectsCount = m_outpuDims[1].d[1];
-
-        const float fw = static_cast<float>(frameSize.width) / static_cast<float>(m_inputDims.d[3]);
-        const float fh = static_cast<float>(frameSize.height) / static_cast<float>(m_inputDims.d[2]);
-
-        //std::cout << "Dets[" << imgIdx << "] = " << dets[imgIdx] << ", objectsCount = " << objectsCount << std::endl;
-
-        const size_t step1 = imgIdx * objectsCount;
-        const size_t step2 = 4 * imgIdx * objectsCount;
-        for (size_t i = 0; i < static_cast<size_t>(dets[imgIdx]); ++i)
-        {
-            // Box
-            const size_t k = i * 4;
-            float class_conf = scores[i + step1];
-            int classId = classes[i + step1];
-            if (class_conf >= m_params.confThreshold)
-            {
-                float x = fw * boxes[k + 0 + step2];
-                float y = fh * boxes[k + 1 + step2];
-                float width = fw * boxes[k + 2 + step2] - x;
-                float height = fh * boxes[k + 3 + step2] - y;
-
-                //if (i == 0)
-                //{
-                //    std::cout << i << ": class_conf = " << class_conf << ", classId = " << classId << " (" << classes[i + step1] << "), rect = " << cv::Rect(cvRound(x), cvRound(y), cvRound(width), cvRound(height)) << std::endl;
-                //    std::cout << "boxes = " << boxes[k + 0 + step2] << ", " << boxes[k + 1 + step2] << ", " << boxes[k + 2 + step2] << ", " << boxes[k + 3 + step2] << std::endl;
-                //}
-                resBoxes.emplace_back(classId, class_conf, cv::Rect(cvRound(x), cvRound(y), cvRound(width), cvRound(height)));
-            }
-        }
-    }
-    else if (outputs.size() == 1)
-    {
-        if (m_params.m_netType == tensor_rt::ModelType::YOLOV8 || m_params.m_netType == tensor_rt::ModelType::YOLOV9)
-        {
-			//0: name: images, size: 1x3x640x640
-		    //1: name: output0, size: 1x84x8400
-
-            const float fw = static_cast<float>(frameSize.width) / static_cast<float>(m_inputDims.d[3]);
-            const float fh = static_cast<float>(frameSize.height) / static_cast<float>(m_inputDims.d[2]);
-
-            auto output = outputs[0];
-
-            size_t ncInd = 1;
-            size_t lenInd = 2;
-            int nc = m_outpuDims[0].d[ncInd] - 4;
-            int dimensions = nc + 4;
-            size_t len = static_cast<size_t>(m_outpuDims[0].d[lenInd]) / m_params.explicitBatchSize;
-            //auto Volume = [](const nvinfer1::Dims& d)
-            //{
-            //    return std::accumulate(d.d, d.d + d.nbDims, 1, std::multiplies<int>());
-            //};
-            auto volume = len * m_outpuDims[0].d[ncInd]; // Volume(m_outpuDims[0]);
-            output += volume * imgIdx;
-            //std::cout << "len = " << len << ", nc = " << nc << ", m_params.confThreshold = " << m_params.confThreshold << ", volume = " << volume << std::endl;
-
-            cv::Mat rawMemory(1, dimensions * static_cast<int>(len), CV_32FC1, output);
-            rawMemory = rawMemory.reshape(1, dimensions);
-            cv::transpose(rawMemory, rawMemory);
-            output = (float*)rawMemory.data;
-
-            //std::cout << "output[0] mem:\n";
-            //for (size_t ii = 0; ii < 100; ++ii)
-            //{
-            //    std::cout << ii << ": ";
-            //    for (size_t jj = 0; jj < 20; ++jj)
-            //    {
-            //        std::cout << output[ii * 20 + jj] << " ";
-            //    }
-            //    std::cout << ";" << std::endl;
-            //}
-            //std::cout << ";" << std::endl;
-
-            std::vector<int> classIds;
-            std::vector<float> confidences;
-            std::vector<cv::Rect> rectBoxes;
-            classIds.reserve(len);
-            confidences.reserve(len);
-            rectBoxes.reserve(len);
-
-			for (size_t i = 0; i < len; ++i)
-			{
-				// Box
-				size_t k = i * (nc + 4);
-
-				int classId = -1;
-				float objectConf = 0.f;
-				for (int j = 0; j < nc; ++j)
-				{
-					const float classConf = output[k + 4 + j];
-					if (classConf > objectConf)
-					{
-						classId = j;
-						objectConf = classConf;
-					}
-				}
-
-				//if (i == 0)
-				//	std::cout << i << ": object_conf = " << object_conf << ", class_conf = " << class_conf << ", classId = " << classId << ", rect = " << cv::Rect(cvRound(x), cvRound(y), cvRound(width), cvRound(height)) << std::endl;
-
-				if (objectConf >= m_params.confThreshold)
-				{
-					classIds.push_back(classId);
-					confidences.push_back(objectConf);
-
-					// (center x, center y, width, height) to (x, y, w, h)
-					float x = fw * (output[k] - output[k + 2] / 2);
-					float y = fh * (output[k + 1] - output[k + 3] / 2);
-					float width = fw * output[k + 2];
-					float height = fh * output[k + 3];
-					rectBoxes.emplace_back(cvRound(x), cvRound(y), cvRound(width), cvRound(height));
-				}
-			}
-
-            // Non-maximum suppression to eliminate redudant overlapping boxes
-            std::vector<int> indices;
-            cv::dnn::NMSBoxes(rectBoxes, confidences, m_params.confThreshold, m_params.nmsThreshold, indices);
-            resBoxes.reserve(indices.size());
-
-            for (size_t bi = 0; bi < indices.size(); ++bi)
-            {
-                resBoxes.emplace_back(classIds[indices[bi]], confidences[indices[bi]], rectBoxes[indices[bi]]);
-            }
-        }
-        else
-        {
-            const float fw = static_cast<float>(frameSize.width) / static_cast<float>(m_inputDims.d[3]);
-            const float fh = static_cast<float>(frameSize.height) / static_cast<float>(m_inputDims.d[2]);
-
-            auto output = outputs[0];
-
-            size_t ncInd = 2;
-            size_t lenInd = 1;
-            if (m_outpuDims[0].nbDims == 2)
-            {
-                ncInd = 1;
-                lenInd = 0;
-            }
-            int nc = m_outpuDims[0].d[ncInd] - 5;
-            size_t len = static_cast<size_t>(m_outpuDims[0].d[lenInd]) / m_params.explicitBatchSize;
-            //auto Volume = [](const nvinfer1::Dims& d)
-            //{
-            //    return std::accumulate(d.d, d.d + d.nbDims, 1, std::multiplies<int>());
-            //};
-            auto volume = len * m_outpuDims[0].d[ncInd]; // Volume(m_outpuDims[0]);
-            output += volume * imgIdx;
-            //std::cout << "len = " << len << ", nc = " << nc << ", m_params.confThreshold = " << m_params.confThreshold << ", volume = " << volume << std::endl;
-
-            if (m_outpuDims[0].nbDims == 2) // With nms
-            {
-                std::vector<int> classIds;
-                std::vector<float> confidences;
-                std::vector<cv::Rect> rectBoxes;
-                classIds.reserve(len);
-                confidences.reserve(len);
-                rectBoxes.reserve(len);
-
-                for (size_t i = 0; i < len; ++i)
-                {
-                    // Box
-                    size_t k = i * 7;
-                    float class_conf = output[k + 6];
-                    int classId = cvRound(output[k + 5]);
-                    if (class_conf >= m_params.confThreshold)
-                    {
-                        float x = fw * output[k + 1];
-                        float y = fh * output[k + 2];
-                        float width = fw * (output[k + 3] - output[k + 1]);
-                        float height = fh * (output[k + 4] - output[k + 2]);
-
-                        //if (i == 0)
-                        //	std::cout << i << ": class_conf = " << class_conf << ", classId = " << classId << ", rect = " << cv::Rect(cvRound(x), cvRound(y), cvRound(width), cvRound(height)) << std::endl;
-
-                        classIds.push_back(classId);
-                        confidences.push_back(class_conf);
-                        rectBoxes.emplace_back(cvRound(x), cvRound(y), cvRound(width), cvRound(height));
-
-                        //bboxes.emplace_back(classId, class_conf, cv::Rect(cvRound(x), cvRound(y), cvRound(width), cvRound(height)));
-                    }
-                }
-
-                // Non-maximum suppression to eliminate redudant overlapping boxes
-                std::vector<int> indices;
-                cv::dnn::NMSBoxes(rectBoxes, confidences, m_params.confThreshold, m_params.nmsThreshold, indices);
-                resBoxes.reserve(indices.size());
-
-                for (size_t bi = 0; bi < indices.size(); ++bi)
-                {
-                    resBoxes.emplace_back(classIds[indices[bi]], confidences[indices[bi]], rectBoxes[indices[bi]]);
-                }
-            }
-            else // Without nms
-            {
-                std::vector<int> classIds;
-                std::vector<float> confidences;
-                std::vector<cv::Rect> rectBoxes;
-                classIds.reserve(len);
-                confidences.reserve(len);
-                rectBoxes.reserve(len);
-
-                for (size_t i = 0; i < len; ++i)
-                {
-                    // Box
-                    size_t k = i * (nc + 5);
-                    float object_conf = output[k + 4];
-
-                    //if (i == 0)
-                    //{
-                    //	std::cout << "mem" << i << ": ";
-                    //	for (size_t ii = 0; ii < nc + 5; ++ii)
-                    //	{
-                    //		std::cout << output[k + ii] << " ";
-                    //	}
-                    //	std::cout << std::endl;
-                    //}
-
-                    if (object_conf >= m_params.confThreshold)
-                    {
-                        // (center x, center y, width, height) to (x, y, w, h)
-                        float x = fw * (output[k] - output[k + 2] / 2);
-                        float y = fh * (output[k + 1] - output[k + 3] / 2);
-                        float width = fw * output[k + 2];
-                        float height = fh * output[k + 3];
-
-                        // Classes
-                        float class_conf = output[k + 5];
-                        int classId = 0;
-
-                        for (int j = 1; j < nc; ++j)
-                        {
-                            if (class_conf < output[k + 5 + j])
-                            {
-                                classId = j;
-                                class_conf = output[k + 5 + j];
-                            }
-                        }
-
-                        class_conf *= object_conf;
-
-                        //if (i == 0)
-                        //	std::cout << i << ": object_conf = " << object_conf << ", class_conf = " << class_conf << ", classId = " << classId << ", rect = " << cv::Rect(cvRound(x), cvRound(y), cvRound(width), cvRound(height)) << std::endl;
-
-                        classIds.push_back(classId);
-                        confidences.push_back(class_conf);
-                        rectBoxes.emplace_back(cvRound(x), cvRound(y), cvRound(width), cvRound(height));
-                    }
-                }
-
-                // Non-maximum suppression to eliminate redudant overlapping boxes
-                std::vector<int> indices;
-                cv::dnn::NMSBoxes(rectBoxes, confidences, m_params.confThreshold, m_params.nmsThreshold, indices);
-                resBoxes.reserve(indices.size());
-
-                for (size_t bi = 0; bi < indices.size(); ++bi)
-                {
-                    resBoxes.emplace_back(classIds[indices[bi]], confidences[indices[bi]], rectBoxes[indices[bi]]);
-                }
-            }
-        }
-    }
-}
